Stop using `Bundler::RubyVersion#to_gem_version_with_patchlevel` (#41)
This method reports the patch level of Ruby, which is no longer useful
since Ruby 2.0.0, and was causing Bundler to show some strange error
messages about Ruby version with 4 segments, like 3.1.1.0.

The method was removed too aggressively from Bundler without knowing
that it was actually being used by bootboot. Depending on how widely
it's actually used, we might restore it with a deprecation, but this
commit makes bootboot use a better method.