Update README references
Primarily, we need to update the README to point to the new version of
the gem (0.2.2).

As a nit and personal opinion, I also updated the upgrade examples.

Although they are valid, I think it is worth updating these every few
years as a sign that this is still being maintained (which it is).

Seeing references to EOL'd versions of Ruby and Rails in 2023 - I figured
I would tweak it while looking at the repo.